+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # NEXT RELEASE
+* Add `invalidTags` property to `TagTootsConfig` type
+* Bump up filtering out of participated tags from favourited tag toots pull to "more than 3 uses"
 
 ### v0.53.1
-* Lower config.batchCompleteSleepBetweenMS from 250ms to 150ms because the completeProperties() seems to be much faster now
+* Lower `config.batchCompleteSleepBetweenMS` from 250ms to 150ms because the `completeProperties()` seems to be much faster now
 
 ## v0.53.0
 * Introduce `TagList` and `TootsForTagsList` classes to manage participated/favourited/trending tags