+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Use a workaround for the weird issue with half width japanese characters
 
 ### v0.53.2
 * Remove japanese/korean/cyrillic etc. tags from trending list unless that's the user's current language